Quality of life manifest generation (#32)
When there is a sha colision, we how write the files to a temp folder
so it is easier to debug.

Also, sorting the manifest makes it easier to diff previous versions
with new versions of the application. This is good for debugging and
could be good for diagnose real production issues on the applications
using mendel, in case they store all manifests with the application
package.