.mendelrc can use nested arrays; flatten on cli; verbose option
This fixes the verbose output by default, making it an option.
It also makes it possible to do nested arrays in .mendelrc files, which is specifically useful for yaml files. Here is a simple example:

```yml
sharedConfig:
  vendorLibs: &vendorLibs
    - react
    - react-dom
    - react-nice-plugin
    - lodash/is/cool
  onDemandBundle: &onDemandBundle
    - ./src/less-used-feature.jsx

bundles:
  main:
    entries:
      - ./src/main.js
    external:
      - *vendorLibs
      - *onDemandBundle
  less-used:
    external: *vendorLibs
    require: *onDemandBundle
```

We are also deprecating use of relative paths on entries with a warn message.