full-example works with mendel-cli again; .mendelrc_v2 details below;
In order to keep the full-example workis (in case anyone checks out the repo to try), I moved the .mendelrc_v1 back to default name and added a _v2 version.

In config, I added a check for MENDELRC environment variable, this way we can force any version we might need, we can even `export MENDELRC=customname` on bash to persist this on our boxes. Lastly I made the cli example on mendel-pipeline to default to `.mendelrc_v2` to make this transparent, run `mendel` and it looks for `.mendelrc` and run `mendel-pipeline <patth>` to and it looks for `.mendelrc_v2`.

@stephanwlee ^