Refactor config; Parse variations; consistent projectRoot.
This commit passes all tests but mendel-pipeline is broken. This is separate from next commit to help understand the steps taken to acomplish the normalizedId implementation.

This also makes sure some configurations that used to be functions are now POJO, since this way we can pass between processes without major concerns about serialization.

Finally, this removes the extra param to pass paths in the demo CLI on mendel-pipeline. Eventually we well replace this by mendel-cli and we can start passing paths again at this version with proper CLI parsing, since nesting base-config with CLI params takes a bit more work.