`normalizedId` now accounts for `package.json`
I feel like this is complete from the calculation side. I looked briefilly to the possibility of invalidating the files instead of throwing and added some inlide comments after I decided it is not time to do it.

There might be implications on what happens when we actually require `./components/about` on full example, as I didn't test this integration as of this commit. Once we do, there are implications for package.json, for instance, we have more than 1 `.components/about` for the `base` variation and we are not disambiguating.

Also, we can optimize the `deps` step to avoid filesystem sometimes. @stephanwlee did you notice how slow is deps in comparison to transforms now that GST is merged? It seems this optimization might be worthwhile after all.