Use shasum instead of integers in prune
As of now, we are using incremented ids in prune post generator.
Order of files cannot be relied on and it causes unexpected
dependecy issues between multiple bundles. This is more vulnerable
when files are added or removed.

We should always create unique hash based on a file name. Irrespective of file order,
prune id should be unique.